+++ packages/terra-time-input/src/terra-dev-site/test/time-input/twelve-hour/Mobile.test.jsx
@@ -69,12 +69,21 @@ export default class TimeInputElement extends React.Component {
           variant={TimeUtil.FORMAT_12_HOUR}
           disabled
         />
-        <h3>Input with Second</h3>
+        <h3>Time Input with Second</h3>
         <TimeInput
           id="timeInputWithSecond"
           name="time-input-with-second"
           variant={TimeUtil.FORMAT_12_HOUR}
           showSeconds
+        />
+        <h3>Time Input with Invalid Meridiem</h3>
+        <TimeInput
+          id="timeInputWithInvalidMeridiem"
+          name="time-input-with-invalid-meridiem"
+          variant={TimeUtil.FORMAT_12_HOUR}
+          showSeconds
+          isInvalid
+          isInvalidMeridiem
         />
       </>
     );